Rename to represent compoenent more accurately
